+++ concordium-cis2/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased changes
+- Add specific parameter type `OnReceivingCis2DataParams<T, A, D>` for a contract function which receives CIS2 tokens with a specific type D for the AdditionalData.
+- Add `Deserial` trait for `OnReceivingCis2DataParams<T, A, D>`.
+- Add `Serial` trait for `OnReceivingCis2DataParams<T, A, D>`.
 
 ## concordium-cis2 5.1.0 (2023-10-18)